requirements.txt
requests==2.28.0
python-dotenv==0.20.0
pyTelegramBotAPI==4.5.1
loguru==0.6.0